move CreateProfileView into App package
